rollouts: addressed Chris's comments
