+++ rollouts/Makefile
@@ -175,7 +175,6 @@ envtest: $(ENVTEST) ## Download envtest-setup locally if necessary.
 $(ENVTEST): $(LOCALBIN)
 	test -s $(LOCALBIN)/setup-envtest || GOBIN=$(LOCALBIN) go install sigs.k8s.io/controller-runtime/tools/setup-envtest@latest
 
-
 .PHONY: github_test
 github_test:
 	go test -v -tags=github ./pkg/packagediscovery/...
